Update systemd, rename some files
