+++ src/gmail2line/__init__.py
@@ -2,5 +2,5 @@
 Main module for Gmail2...
 """
 from gmail2line.cli import exit_codes
-from gmail2line.core.patterns import find_matches
+from gmail2line.core.matching import find_matches
 from gmail2line.gmail import mail, label, resource
