+++ src/gmail2line/core/matching.py
